+++ adminclient/src/Helpers/Courses.js
@@ -1,37 +1,54 @@
-export const courseData = [
-        {
-          title: "HR AND HIRING",
-          catogories: [
-            {
-              id: "16",
-              image: `https://cdn.pixabay.com/photo/2015/04/23/22/00/tree-736885__480.jpg`,
-              subject: "biology",
-            },
-            {
-              id: "15",
-              image: `https://cdn.pixabay.com/photo/2015/04/23/22/00/tree-736885__480.jpg`,
-              subject: "biology",
-            },
-            {
-              id: "13",
-              image: `https://cdn.pixabay.com/photo/2015/04/23/22/00/tree-736885__480.jpg`,
-              subject: "biology",
-            },
-          ],
-        },
-        {
-          title: "Tech Courses",
-          catogories: [
-            {
-              id: "10",
-              image: `https://cdn.pixabay.com/photo/2015/04/23/22/00/tree-736885__480.jpg`,
-              subject: "biology",
-            },
-            {
-              id: "11",
-              image: `https://cdn.pixabay.com/photo/2015/04/23/22/00/tree-736885__480.jpg`,
-              subject: "biology",
-            },
-          ],
-        },
-      ];
+const courses = [
+  {
+    title: "HR AND HIRING",
+    catogories: [
+      {
+        id: "16",
+        image: `https://cdn.pixabay.com/photo/2015/04/23/22/00/tree-736885__480.jpg`,
+        subject: "biology",
+      },
+      {
+        id: "15",
+        image: `https://cdn.pixabay.com/photo/2015/04/23/22/00/tree-736885__480.jpg`,
+        subject: "biology",
+      },
+      {
+        id: "13",
+        image: `https://cdn.pixabay.com/photo/2015/04/23/22/00/tree-736885__480.jpg`,
+        subject: "biology",
+      },
+      {
+        id: "15",
+        image: `https://cdn.pixabay.com/photo/2015/04/23/22/00/tree-736885__480.jpg`,
+        subject: "biology",
+      },
+      {
+        id: "15",
+        image: `https://cdn.pixabay.com/photo/2015/04/23/22/00/tree-736885__480.jpg`,
+        subject: "biology",
+      },
+      {
+        id: "15",
+        image: `https://cdn.pixabay.com/photo/2015/04/23/22/00/tree-736885__480.jpg`,
+        subject: "biology",
+      },
+    ],
+  },
+  {
+    title: "TECH COURSES",
+    catogories: [
+      {
+        id: "10",
+        image: `https://cdn.pixabay.com/photo/2015/04/23/22/00/tree-736885__480.jpg`,
+        subject: "biology",
+      },
+      {
+        id: "11",
+        image: `https://cdn.pixabay.com/photo/2015/04/23/22/00/tree-736885__480.jpg`,
+        subject: "biology",
+      },
+    ],
+  },
+];
+
+export default courses;
